Backend/requirements.txt
 Django==5.2.8
 djangorestframework
 djoser
 django-filter
 django-cors-headers
 psycopg2-binary
 gunicorn
 dj-database-url
